fix variable scoping / resolution problem (ch 11)
The book describes how to resolve variables after parse-time but before
the code is executed. There's a problem with the implementation of
environments _before_ this chapter, which is illustrated by the
following program:

```
var a = "global";
{
  fun showA() {
    print a;
  }

  showA();

  var a = "block";
  showA();
}
```

With the chapter 10 implementation it will print

```
"global"
"block"
```

this is because once we're inside of the block that starts on line 2
we'll start a new environment. The environment inside of the `showA`
function will then have this environment as a parent, and when it does
the recursive lookup to find the value of `a` it will check first this
block scope and then the outer global scope. All well and good the first
time that `showA` is called, but this will produce incorrect behavior
after the second call since now there is a different variable `a`
defined on the block-associated `Environment`.

The book's solution is to add a variable resolution pass which occurs
after the AST is parsed. This basically does analysis on every variable
declaration _or_ reassignment _or_ access and figures out in each place
how many 'hops' on environments up the chain the concrete action being
done with that variable should traverse before it actually happens. Thus
a variable with the same name being bound on an intermediate environment
will no longer result in incorrect behavior.

To me this seems to make things more complicated than they need to be,
especially since the book's implementation uses the Visitor pattern
which I don't really like.

Instead, I made a few changes to basically create a new environment
every time a variable is declared. This required changing the
implementation of my `interpret` function a little bit so that it will
always return an environment - for most operations this is the same
environment that was passed in, but for variable declarations it is
instead a new environment, with the environment passed in as a parent,
which is then passed in to the recursive function which actually
interprets the variable declaration. Then code can be interpreted by
basically reducing across a list of statements, starting with some
initial environment. This looks something like:

```ts
function interpret(statements: Stmt[], env: Environment): Environment {
   ...
   return statements.reduce(execute, env)
}

function execute(env: Environment, statement: Stmt): Environment {
  ...
}
```

`interpret` returns an `Environment` to make it easy to have a
persistent environment between lines entered into the repl, but when
just executing a script that returned object won't be used for anything.
This fixes the variable resolution / scoping problem in a simple and
effective way and is, I think, more elegant than what the book does.